feat(build): 添加 wasmJs 平台支持，并合并部分 JS 和 wasmJs 到 web
