+++ simbot-component-kook/simbot-component-kook-stdlib/build.gradle.kts
@@ -1,4 +1,29 @@
 /*
+ *     Copyright (c) 2026. ForteScarlet.
+ *
+ *     Project    https://github.com/simple-robot/simpler-robot
+ *     Email      ForteScarlet@163.com
+ *
+ *     This file is part of the Simple Robot Library (Alias: simple-robot, simbot, etc.).
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU Lesser General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     Lesser GNU General Public License for more details.
+ *
+ *     You should have received a copy of the Lesser GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ */
+
+import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
+
+/*
  *     Copyright (c) 2023-2026. ForteScarlet.
  *
  *     Project    https://github.com/simple-robot/simpler-robot
@@ -42,6 +67,11 @@ kotlin {
         configJs()
     }
 
+    @OptIn(ExperimentalWasmDsl::class)
+    wasmJs {
+        configWasmJs()
+    }
+
     applyTier123(supportKtorClient = true)
 
 
@@ -66,8 +96,16 @@ kotlin {
             implementation(project(":simbot-logger-slf4j2-impl"))
         }
 
+        webMain.dependencies {
+            api(kotlinWrappers.web)
+        }
+
         jsMain.dependencies {
             api(libs.ktor.client.js)
+        }
+
+        wasmJsMain.dependencies {
+            api(libs.ktor.client.cio)
         }
 
         mingwTest.dependencies {
