+++ simbot-component-kook/simbot-component-kook-stdlib/src/jsMain/kotlin/love/forte/simbot/kook/stdlib/internal/BotStates.js.kt
@@ -24,48 +24,48 @@
 package love.forte.simbot.kook.stdlib.internal
 
 import io.ktor.websocket.*
-import kotlinx.coroutines.await
-import love.forte.simbot.annotations.InternalSimbotAPI
-import org.khronos.webgl.Uint8Array
-import org.w3c.files.Blob
-import kotlin.js.Promise
+import js.buffer.toArrayBuffer
+import web.blob.Blob
 
-/**
- * 由平台实现对二进制 `deflate` 压缩数据进行解压缩并转为字符串数据。
- *
- * 使用 [DecompressionStream](https://nodejs.org/api/webstreams.html#class-decompressionstream)
- * 对数据进行解压缩、并最终使用 [TextDecoder](https://developer.mozilla.org/en-US/docs/Web/API/TextDecoder) 解析字符串。
- *
- * _Note: JS 平台下的二进制数据是**实验性**的，不保证可用性，且在未来可能会修改/删除。_
- */
-@InternalSimbotAPI
-public actual suspend fun Frame.Binary.readToTextWithDeflated(): String {
-    val uint8Array0 = Uint8Array(data.toTypedArray())
+@OptIn(ExperimentalWasmJsInterop::class)
+internal actual fun Frame.Binary.dataToBlob(): Blob = Blob(arrayOf(data.toArrayBuffer()).toJsArray())
 
-    @Suppress("UNUSED_VARIABLE")
-    val blob = Blob(arrayOf(uint8Array0))
-
-    val decompressionStream = js("new DecompressionStream('deflate')")
-    val decompressedStream = js("blob.stream()").pipeThrough(decompressionStream)
-
-    val reader = decompressedStream.getReader()
-
-    val chunks = mutableListOf<Uint8Array>()
-    var result = (reader.read() as Promise<dynamic>).await()
-
-    while (!result.done as Boolean) {
-        val rv = result.value
-        @Suppress("UnsafeCastFromDynamic")
-        chunks.add(rv)
-        result = (reader.read() as Promise<dynamic>).await()
-    }
-
-    val uint8Array = Uint8Array(chunks.sumOf { it.length })
-    var offset = 0
-    chunks.forEach {
-        uint8Array.set(it, offset)
-        offset += it.length
-    }
-
-    return js("new TextDecoder()").decode(uint8Array) as String
-}
+// /**
+//  * 由平台实现对二进制 `deflate` 压缩数据进行解压缩并转为字符串数据。
+//  *
+//  * 使用 [DecompressionStream](https://nodejs.org/api/webstreams.html#class-decompressionstream)
+//  * 对数据进行解压缩、并最终使用 [TextDecoder](https://developer.mozilla.org/en-US/docs/Web/API/TextDecoder) 解析字符串。
+//  *
+//  * _Note: JS 平台下的二进制数据是**实验性**的，不保证可用性，且在未来可能会修改/删除。_
+//  */
+// @InternalSimbotAPI
+// public actual suspend fun Frame.Binary.readToTextWithDeflated(): String {
+//     val uint8Array0 = Uint8Array(data.toTypedArray())
+//
+//     @Suppress("UNUSED_VARIABLE")
+//     val blob = io.ktor.client.fetch.Blob(arrayOf(uint8Array0))
+//
+//     val decompressionStream = js("new DecompressionStream('deflate')")
+//     val decompressedStream = js("blob.stream()").pipeThrough(decompressionStream)
+//
+//     val reader = decompressedStream.getReader()
+//
+//     val chunks = mutableListOf<Uint8Array>()
+//     var result = (reader.read() as Promise<dynamic>).await()
+//
+//     while (!result.done as Boolean) {
+//         val rv = result.value
+//         @Suppress("UnsafeCastFromDynamic")
+//         chunks.add(rv)
+//         result = (reader.read() as Promise<dynamic>).await()
+//     }
+//
+//     val uint8Array = Uint8Array(chunks.sumOf { it.length })
+//     var offset = 0
+//     chunks.forEach {
+//         uint8Array.set(it, offset)
+//         offset += it.length
+//     }
+//
+//     return js("new TextDecoder()").decode(uint8Array) as String
+// }
