+++ simbot-component-onebot/simbot-component-onebot-v11/simbot-component-onebot-v11-core/src/webMain/kotlin/love/forte/simbot/component/onebot/v11/core/component/OneBot11Component.web.kt
