+++ simbot-component-onebot/simbot-component-onebot-v11/simbot-component-onebot-v11-event/build.gradle.kts
@@ -23,6 +23,7 @@
 
 import com.google.devtools.ksp.gradle.KspAATask
 import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
+import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
 
 plugins {
     id("simbot.kotlin-multiplatform")
@@ -50,6 +51,11 @@ kotlin {
 
     js(IR) {
         configJs()
+    }
+
+    @OptIn(ExperimentalWasmDsl::class)
+    wasmJs {
+        configWasmJs()
     }
 
     applyTier123()
