+++ simbot-component-onebot/simbot-component-onebot-v11/simbot-component-onebot-v11-message/src/webMain/kotlin/love/forte/simbot/component/onebot/v11/message/MessageElementResolvers.web.kt
