+++ simbot-component-qq/simbot-component-qq-guild-api/build.gradle.kts
@@ -22,6 +22,7 @@
  */
 
 import com.google.devtools.ksp.gradle.KspAATask
+import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
 
 plugins {
     id("simbot.kotlin-multiplatform")
@@ -49,6 +50,11 @@ kotlin {
     }
 
     applyTier123(supportKtorClient = true)
+
+    @OptIn(ExperimentalWasmDsl::class)
+    wasmJs {
+        configWasmJs()
+    }
 
     sourceSets {
         commonMain {
