+++ simbot-component-qq/simbot-component-qq-guild-api/src/webMain/kotlin/love/forte/simbot/qguild/api/message/resolveOther.web.kt
@@ -1,0 +1,34 @@
+/*
+ *     Copyright (c) 2023-2026. ForteScarlet.
+ *
+ *     Project    https://github.com/simple-robot/simpler-robot
+ *     Email      ForteScarlet@163.com
+ *
+ *     This file is part of the Simple Robot Library (Alias: simple-robot, simbot, etc.).
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU Lesser General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     Lesser GNU General Public License for more details.
+ *
+ *     You should have received a copy of the Lesser GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ */
+
+package love.forte.simbot.qguild.api.message
+
+import io.ktor.client.request.forms.*
+import love.forte.simbot.qguild.QGInternalApi
+
+/**
+ * Web 平台所能支持的针对 [MessageSendApi.Body.fileImage] 的解析: 无额外支持。
+ *
+ */
+@QGInternalApi
+public actual fun FormBuilder.resolveOther(fileImage: Any?): Unit = Unit
