--- simbot-component-qq/simbot-component-qq-guild-stdlib/src/jsMain/kotlin/love/forte/simbot/qguild/stdlib/BotRequests.js.kt
@@ -1,55 +1,0 @@
-/*
- * Copyright (c) 2023-2024. ForteScarlet.
- *
- * This file is part of simbot-component-qq-guild.
- *
- * simbot-component-qq-guild is free software: you can redistribute it and/or modify it under the terms
- * of the GNU Lesser General Public License as published by the Free Software Foundation,
- * either version 3 of the License, or (at your option) any later version.
- *
- * simbot-component-qq-guild is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License along with simbot-component-qq-guild.
- * If not, see <https://www.gnu.org/licenses/>.
- */
-
-package love.forte.simbot.qguild.stdlib
-
-import io.ktor.client.statement.*
-import kotlinx.coroutines.promise
-import love.forte.simbot.annotations.Api4Js
-import love.forte.simbot.qguild.api.QQGuildApi
-import kotlin.js.Promise
-
-
-/**
- * 直接通过bot进行请求。
- *
- * @throws love.forte.simbot.qguild.QQGuildApiException 如果返回状态码不在 200..300之间。
- */
-@Api4Js
-public fun Bot.requestByAsync(api: QQGuildApi<*>): Promise<HttpResponse> = promise {
-    request(api)
-}
-
-/**
- * 直接通过bot进行请求。
- *
- * @throws love.forte.simbot.qguild.QQGuildApiException 如果返回状态码不在 200..300之间。
- */
-@Api4Js
-public fun Bot.requestTextByAsync(api: QQGuildApi<*>): Promise<String> = promise {
-    requestText(api)
-}
-
-/**
- * 直接通过bot进行请求。
- *
- * @throws love.forte.simbot.qguild.QQGuildApiException 如果返回状态码不在 200..300之间。
- */
-@Api4Js
-public fun <R : Any> Bot.requestDataByAsync(api: QQGuildApi<R>): Promise<R> = promise {
-    requestData(api)
-}
